modularised the code
