+++ Data Engineering/ecommerce_pipeline/airflow/airflow.cfg
@@ -209,7 +209,7 @@ unit_test_mode = False
 #
 # Variable: AIRFLOW__CORE__ENABLE_XCOM_PICKLING
 #
-enable_xcom_pickling = False
+enable_xcom_pickling = True
 
 # What classes can be imported during deserialization. This is a multi line value.
 # The individual items will be parsed as a pattern to a glob function.
