+++ Data Engineering/ecommerce_pipeline/airflow/dags/dag__transactions.py
@@ -1,36 +1,81 @@
+## ------------ Imports ------------ ##
 from datetime import datetime, timedelta
 
 from airflow.models import DAG
 from airflow.operators.python import PythonOperator
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
 
-import python_files.transactions as transactions
+import python_files.faker_transactions as faker_transactions
+import python_files.convert_dataclass_to_json as convert_dataclass_to_json
+import python_files.export_list_as_json as export_list_as_json
+import python_files.export_to_s3 as export_to_s3
 
 
 
+## ------------ Defaults/Variables ------------ ##
 default_args = {
     "owner": "waseem",
     "retries": 1,
     "retry_delay":timedelta(minutes=5)
 }
 
+TRANSACTION_MIN = 1
+TRANSACTION_MAX = 9
+LOCAL_FILE = "transactions.json"
+NAME_FOR_S3_FILE = 'transactions_s3.json'
 
 
-def task_1():
-    transactions.get_transactions_load_s3()
+
+## ------------ Python Callables ------------ ##
+def task__get_transactions():
+    transactions = faker_transactions.get_transactions(TRANSACTION_MIN, TRANSACTION_MAX)
+    return transactions
 
 
+def task__dataclass_to_json(ti):
+    transactions = convert_dataclass_to_json.\
+                    dataclass_to_json(dataclass_list=ti.xcom_pull(task_ids='get_transactions'))
+    return transactions
+
+
+def task__list_to_json_file(ti):
+    export_list_as_json.list_to_json_file(lst=ti.xcom_pull(task_ids='dataclass_to_json'),
+                                          file_name=LOCAL_FILE)
+
+
+def task__save_json_file_to_s3(ti):
+    export_to_s3.save_json_file_to_s3(local_file=LOCAL_FILE, 
+                                      file_name=NAME_FOR_S3_FILE)
+
+
+
+## ------------ Dags & Tasks ------------ ##
 with DAG(
     default_args=default_args,
     dag_id="transactions",
-    description="extracts transactions and loads to s3",
+    description="Pipeline",
     start_date=datetime(2024,12,29),
     schedule_interval="@daily"
 ) as dag:
 
-    task1 = PythonOperator(
-        task_id="get_transactions_load_s3",
-        python_callable=task_1
+    get_transactions = PythonOperator(
+        task_id="get_transactions",
+        python_callable=task__get_transactions
     )
 
-    task1
+    dataclass_to_json = PythonOperator(
+        task_id="dataclass_to_json",
+        python_callable=task__dataclass_to_json
+    )
+
+    list_to_json_file = PythonOperator(
+        task_id="list_to_json_file",
+        python_callable=task__list_to_json_file
+    )
+
+    save_json_file_to_s3 = PythonOperator(
+        task_id="save_json_file_to_s3",
+        python_callable=task__save_json_file_to_s3
+    )
+
+    get_transactions>>dataclass_to_json>>list_to_json_file>>save_json_file_to_s3
